feat(Nav): separate the interfaces composing the props navigation
