+++ src/components/navigation/Nav/interfaces/Nav.LinkComponet.interface.ts
@@ -1,0 +1,5 @@
+import { Link } from "./Nav.Link.interface";
+
+export interface INavLinkProps {
+  section: Link[];
+}
