+++ src/components/navigation/Nav/interfaces/Nav.Navigation.interface.ts
@@ -1,0 +1,6 @@
+import { Section } from "./Nav.Sections.interface";
+
+export interface Navigation {
+  title: string;
+  sections: { [key: string]: Section };
+}
