+++ src/components/navigation/Nav/interfaces/Nav.Sections.interface.ts
@@ -1,0 +1,6 @@
+import { Link } from "./Nav.Link.interface";
+
+export interface Section {
+  name: string;
+  links: { [key: string]: Link };
+}
