+++ src/components/navigation/Nav/interfaces/Nav.interface.ts
@@ -1,19 +1,4 @@
-export interface Link {
-  id: string;
-  label: string;
-  icon: Element;
-  path: string;
-}
-
-export interface Section {
-  name: string;
-  links: { [key: string]: Link };
-}
-
-export interface Navigation {
-  title: string;
-  sections: { [key: string]: Section };
-}
+import { Navigation } from "./Nav.Navigation.interface";
 
 export interface INavProps {
   navigation: Navigation;
